[ANDROSDK-1386] Force RelativePeriods adapter to return an Enum
